Can not implement car selection that client own in Deal Create Form. Also minor changes

+++ app/Http/Controllers/DealController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Deal;
+use App\Models\Client;
 use Illuminate\Http\Request;
 
 class DealController extends Controller
@@ -19,9 +20,18 @@ class DealController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
-        return view('deals.create');
+        $clients = Client::all();
+        $selectedClient = null;
+
+        if($request->has('client_id')) {
+            $selectedClient = Client::find($request->client_id);
+            // return redirect()->route('deals.create')->with('selected_client', $selectedClient)->withInput();
+            session(['selected_client' => $selectedClient]);
+        }
+
+        return view('deals.create', compact('clients', 'selectedClient'));
     }
 
     /**
@@ -30,8 +40,18 @@ class DealController extends Controller
     public function store(Request $request)
     {
         $validated = $request->validate([
-
+            'date' => 'required|date|after:today',
+            'client_id' => 'required|exists:clients,id',
+            'service_name' => 'required|string|max:255',
+            'price' => 'required|numeric|min:0',
+            'comment' => 'nullable|string|max:255',
         ]);
+
+        $validated['status'] = 'new';
+
+        Deal::create($validated);
+
+        return redirect()->route('deals.index')->with('success', 'Deal created');
     }
 
     /**
@@ -63,8 +83,10 @@ class DealController extends Controller
      */
     public function destroy(Deal $deal)
     {
+        if (!$deal->exists()) {
+            return redirect()->route('deals.index')->with('error', 'Deal was not deleted');
+        }
         $deal->delete();
-
         return redirect()->route('deals.index')->with('success', 'Deal was deleted');
     }
 }
